+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Speed up shell startup by preferring `HOMEBREW_REPOSITORY` over the slow `brew --repo` call ([#35](https://github.com/tmatilai/chtf/issues/35))
 * Support sourcing and using chtf with `set -u` enabled in bash and zsh ([#35](https://github.com/tmatilai/chtf/issues/35))
+* Fix listing of prerelease versions installed as Casks, e.g. `0.10.0-rc1` ([#36](https://github.com/tmatilai/chtf/issues/36))
 
 
 ## 2.4.0 / 2026-09-07