+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 2.5.1 / _Not released yet_
+
+
 ## 2.5.0 / 2026-09-11
 
 ### Features